First version of Watch app
