+++ Stengt tunnel watch Watch App/Stengt_tunnel_watchApp.swift
@@ -1,0 +1,17 @@
+//
+//  Stengt_tunnel_watchApp.swift
+//  Stengt tunnel watch Watch App
+//
+//  Created by Ole-Kenneth on 30/08/2024.
+//
+
+import SwiftUI
+
+@main
+struct Stengt_tunnel_watch_Watch_AppApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}
