+++ Stengttunnel.xcodeproj/xcuserdata/olekenneth.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -9,6 +9,11 @@
 			<key>orderHint</key>
 			<integer>1</integer>
 		</dict>
+		<key>Stengt tunnel watch Watch App.xcscheme_^#shared#^_</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>1</integer>
+		</dict>
 		<key>Stengt tunnel.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
@@ -17,6 +22,11 @@
 	</dict>
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
+		<key>B0649EA52C825424001E648E</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
 		<key>B0C85D022A7A4FF1004EC4B8</key>
 		<dict>
 			<key>primary</key>
