+++ stengttunnel/Assets.xcassets/App.imageset/Contents.json
@@ -14,6 +14,11 @@
       "filename" : "original_original_app-icon-v4-exportready.png",
       "idiom" : "universal",
       "scale" : "3x"
+    },
+    {
+      "filename" : "original_original_app-icon-v4-exportready 3.png",
+      "idiom" : "watch",
+      "scale" : "2x"
     }
   ],
   "info" : {
