+++ stengttunnel/Assets.xcassets/green.colorset/Contents.json
@@ -29,6 +29,18 @@
         }
       },
       "idiom" : "universal"
+    },
+    {
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "0.596",
+          "green" : "0.757",
+          "red" : "0.424"
+        }
+      },
+      "idiom" : "watch"
     }
   ],
   "info" : {
