+++ stengttunnel/Assets.xcassets/lightGray.colorset/Contents.json
@@ -24,6 +24,18 @@
         }
       },
       "idiom" : "universal"
+    },
+    {
+      "color" : {
+        "color-space" : "extended-srgb",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "0.918",
+          "green" : "0.898",
+          "red" : "0.898"
+        }
+      },
+      "idiom" : "watch"
     }
   ],
   "info" : {
