+++ stengttunnel/Assets.xcassets/yellow.colorset/Contents.json
@@ -29,6 +29,18 @@
         }
       },
       "idiom" : "universal"
+    },
+    {
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "51",
+          "green" : "168",
+          "red" : "251"
+        }
+      },
+      "idiom" : "watch"
     }
   ],
   "info" : {
